Cambio de tipos de retorno a decimal y nuevo componente Razor
Se ha modificado la interfaz `ICartService` en el archivo `ICartService.cs` para cambiar el tipo de retorno del método `GetTotalAmountCartAsync` de `int` a `decimal`.

En el archivo `CartRepository.cs`, se ha cambiado el tipo de retorno del método `GetTotalAmountAsync` de `int` a `decimal`.

En la interfaz `ICartRepository` en el archivo `ICartRepository.cs`, se ha cambiado el tipo de retorno del método `GetTotalAmountAsync` de `int` a `decimal`.

En el archivo `CartServices.cs`, se ha cambiado el tipo de retorno del método `GetTotalAmountCartAsync` de `int` a `decimal`.

Se ha añadido un nuevo componente Razor `CartPeople.razor` que utiliza varias inyecciones de dependencias y muestra un resumen de la compra. Este componente también maneja la inicialización del estado del carrito y la obtención del monto total del carrito como un `decimal`.

+++ Api/Interfaces/ICartService.cs
@@ -7,6 +7,6 @@ namespace EccomercePage.Api.Interfaces
         Task<int> GetTotalProductInCartAsync(string userId);
         Task<CartResumeDTO> GetCartResumeAsync(string userId);
         Task<bool> ChangeItemQuantity(ChangeItemQuantityDTO dto);
-        Task<int> GetTotalAmountCartAsync(string userId);
+        Task<decimal> GetTotalAmountCartAsync(string userId);
     }
 }
